refactor: opanAPI 파싱 로직 유틸 분리
